backend/requirements.txt
fastapi
SQLAlchemy
uvicorn
httpx
PyYAML
python-jose
pycryptodome
python-dotenv
python-multipart
pytest
httpx_auth
requests
psycopg2